Step R35: ManyToManyTracking moved to Tier B, and it found a real defect (#56)
R16 examined this move and deferred it. What makes it worth taking is R13a's lesson: the move is
what makes the tests real. 200/0/200 -> 200/1/201.

The UseTransaction override is written in the same commit as the store switch, and this is the
case CLAUDE.md D6 is about rather than a formality: the core base routes 47 call sites through
ExecuteWithStrategyInTransactionAsync, each opening one transaction every other context must
enlist in. On Tier A it was ignored; here it is real, and the run completes in six seconds
instead of producing D6's lock-timeout shape.

Two things the Tier A class asserted about itself are deleted, because the move makes them
false. The ExecuteWithStrategyInTransactionAsync reseed override said "without a real
transaction there is no rollback to undo the test's mutations" -- true of InMemory, false here.
And SupportsDatabaseDefaults => false said "the backend is the InMemory store, which has no
database default values"; the fixture now declares the six HasDefaultValue/HasDefaultValueSql
statements EF's own SQLite fixture declares. Those are modelling APIs, so both models make the
statement for themselves out of one OnModelCreating -- not the B4/B6/B12 hazard.

The relational base's one added test, Many_to_many_delete_behaviors_are_set, passes.

BROKEN 1: Can_delete_with_many_to_many, and it is a real defect of this provider that only a real
store could expose. It deletes an EntityOne and an EntityTwo whose JoinOneToTwo rows are
cascade-deleted, and the server reports SQLite Error 19: 'FOREIGN KEY constraint failed'. The
join rows must be deleted before their principals -- a relational provider's CommandBatchPreparer
topologically sorts for exactly this -- and something in the wire round trip does not preserve
that order. InMemory enforces no foreign key, so on Tier A this passed while saying nothing.

Not convergence, and that was checked rather than assumed: ManyToManyTrackingSqliteTest declares
zero test overrides, so EF's own SQLite run passes it. Left failing per ADR-004, and worth an
issue of its own -- same class of find as Phase U, which R19 turned up the same way.

failed 78 -> 79, total 27310 -> 27311. Compliance missing bases 57 -> 56, fixtures 19 -> 18.
