+++ test/known-failures.names.txt
@@ -12,6 +12,7 @@ InfoCarrier.Core.FunctionalTests.Sqlite.CustomConvertersInfoCarrierTest.Composit
 InfoCarrier.Core.FunctionalTests.Sqlite.DataAnnotationInfoCarrierTest.Table_can_configure_TPT_with_Owned
 InfoCarrier.Core.FunctionalTests.Sqlite.EntitySplittingInfoCarrierTest.ExecuteDelete_throws_for_entity_splitting(async: False)
 InfoCarrier.Core.FunctionalTests.Sqlite.EntitySplittingInfoCarrierTest.ExecuteDelete_throws_for_entity_splitting(async: True)
+InfoCarrier.Core.FunctionalTests.Sqlite.ManyToManyTrackingInfoCarrierTest.Can_delete_with_many_to_many
 InfoCarrier.Core.FunctionalTests.Sqlite.Query.Associations.OwnedJsonStructuralEqualityQueryInfoCarrierTest.Associate_with_parameter_null
 InfoCarrier.Core.FunctionalTests.Sqlite.Query.Associations.OwnedJsonStructuralEqualityQueryInfoCarrierTest.Contains_with_nested_and_composed_operators
 InfoCarrier.Core.FunctionalTests.Sqlite.Query.Associations.OwnedJsonStructuralEqualityQueryInfoCarrierTest.Contains_with_operators_composed_on_the_collection
